Add instructions test.
